Spec 0013 — Move Base/*Manifest concrete → Manifests/
AttributeManifest, ModelManifest and ShippingManifest land at the new
top-level Manifests/ namespace. Their interface contracts stay in place
for now (covered by the contracts category).
