+++ packages/core/src/LunarServiceProvider.php
@@ -18,18 +18,15 @@ use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
 use Lunar\Core\Addons\Manifest;
 use Lunar\Core\Auth\Manifest as AccessControlManifest;
-use Lunar\Core\Base\AttributeManifest;
 use Lunar\Core\Base\AttributeManifestInterface;
 use Lunar\Core\Base\CartSessionInterface;
 use Lunar\Core\Base\DiscountManagerInterface;
-use Lunar\Core\Base\ModelManifest;
 use Lunar\Core\Base\ModelManifestInterface;
 use Lunar\Core\Base\OrderReferenceGenerator;
 use Lunar\Core\Base\OrderReferenceGeneratorInterface;
 use Lunar\Core\Base\PaymentManagerInterface;
 use Lunar\Core\Base\PricingManagerInterface;
 use Lunar\Core\Base\ProvidesTelemetryInsights;
-use Lunar\Core\Base\ShippingManifest;
 use Lunar\Core\Base\ShippingManifestInterface;
 use Lunar\Core\Base\StorefrontSessionInterface;
 use Lunar\Core\Base\TaxManagerInterface;
@@ -55,6 +52,9 @@ use Lunar\Core\Managers\PaymentManager;
 use Lunar\Core\Managers\PricingManager;
 use Lunar\Core\Managers\StorefrontSessionManager;
 use Lunar\Core\Managers\TaxManager;
+use Lunar\Core\Manifests\AttributeManifest;
+use Lunar\Core\Manifests\ModelManifest;
+use Lunar\Core\Manifests\ShippingManifest;
 use Lunar\Core\Models\Address;
 use Lunar\Core\Models\CartLine;
 use Lunar\Core\Models\Channel;
