+++ packages/upgrade/src/Rector/LunarSetList.php
@@ -44,7 +44,7 @@ final class LunarSetList
         'Lunar\\Actions\\Taxes\\GetTaxZoneState' => 'Lunar\\Core\\Actions\\Taxes\\GetTaxZoneState',
         'Lunar\\Addons\\Manifest' => 'Lunar\\Core\\Addons\\Manifest',
         'Lunar\\Base\\Addressable' => 'Lunar\\Core\\Base\\Addressable',
-        'Lunar\\Base\\AttributeManifest' => 'Lunar\\Core\\Base\\AttributeManifest',
+        'Lunar\\Base\\AttributeManifest' => 'Lunar\\Core\\Manifests\\AttributeManifest',
         'Lunar\\Base\\AttributeManifestInterface' => 'Lunar\\Core\\Base\\AttributeManifestInterface',
         'Lunar\\Base\\BaseModel' => 'Lunar\\Core\\Base\\BaseModel',
         'Lunar\\Base\\CartLineModifier' => 'Lunar\\Core\\Modifiers\\CartLineModifier',
@@ -76,7 +76,7 @@ final class LunarSetList
         'Lunar\\Base\\LunarUser' => 'Lunar\\Core\\Base\\LunarUser',
         'Lunar\\Base\\MediaDefinitionsInterface' => 'Lunar\\Core\\Base\\MediaDefinitionsInterface',
         'Lunar\\Base\\Migration' => 'Lunar\\Core\\Base\\Migration',
-        'Lunar\\Base\\ModelManifest' => 'Lunar\\Core\\Base\\ModelManifest',
+        'Lunar\\Base\\ModelManifest' => 'Lunar\\Core\\Manifests\\ModelManifest',
         'Lunar\\Base\\ModelManifestInterface' => 'Lunar\\Core\\Base\\ModelManifestInterface',
         'Lunar\\Base\\OrderModifier' => 'Lunar\\Core\\Modifiers\\OrderModifier',
         'Lunar\\Base\\OrderModifiers' => 'Lunar\\Core\\Modifiers\\OrderModifiers',
@@ -87,7 +87,7 @@ final class LunarSetList
         'Lunar\\Base\\PricingManagerInterface' => 'Lunar\\Core\\Base\\PricingManagerInterface',
         'Lunar\\Base\\ProvidesTelemetryInsights' => 'Lunar\\Core\\Base\\ProvidesTelemetryInsights',
         'Lunar\\Base\\Purchasable' => 'Lunar\\Core\\Base\\Purchasable',
-        'Lunar\\Base\\ShippingManifest' => 'Lunar\\Core\\Base\\ShippingManifest',
+        'Lunar\\Base\\ShippingManifest' => 'Lunar\\Core\\Manifests\\ShippingManifest',
         'Lunar\\Base\\ShippingManifestInterface' => 'Lunar\\Core\\Base\\ShippingManifestInterface',
         'Lunar\\Base\\ShippingModifier' => 'Lunar\\Core\\Modifiers\\ShippingModifier',
         'Lunar\\Base\\ShippingModifiers' => 'Lunar\\Core\\Modifiers\\ShippingModifiers',
